requirement.txt
flask==2.3.2
flask_login==0.6.2
Flask-WTF==1.1.1
flask_sqlalchemy==3.0.5
loguru==0.6.0
pymysql==1.1.0
flask_mysqldb==1.0.1